+++ src/main/java/gov/nih/nci/evs/api/service/SparqlQueryManagerServiceImpl.java
@@ -205,6 +205,13 @@ public class SparqlQueryManagerServiceImpl implements SparqlQueryManagerService 
       term.setSource(b.getSource().getValue());
       term.setTerminology(getTerm(term.getSource()));
 
+      // Special handling for chebi to avoid "247.0" as a version
+      // We are doing this to avoid a much more complicated sparql-queries.properties file
+      // but may want to do something about this going forward.
+      if (term.getTerminology().equals("chebi")) {
+        term.setVersion(term.getVersion().replaceFirst(".0$", ""));
+      }
+
       final String startDate =
           FhirUtility.convertToYYYYMMDD(
               (b.getDate() == null) ? term.getVersion() : b.getDate().getValue());
@@ -2039,6 +2046,31 @@ public class SparqlQueryManagerServiceImpl implements SparqlQueryManagerService 
     final Map<String, String> values =
         ConceptUtils.asMap("propertyCode", propertyCode, "namedGraph", terminology.getGraph());
     final String query = queryBuilderService.constructQuery("axiom.qualifier", terminology, values);
+    final String res = restUtils.runSPARQL(queryPrefix + query, getQueryURL());
+
+    final ObjectMapper mapper = new ObjectMapper();
+    mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+    final List<String> propertyValues = new ArrayList<>();
+
+    final Sparql sparqlResult = mapper.readValue(res, Sparql.class);
+    final Bindings[] bindings = sparqlResult.getResults().getBindings();
+    for (final Bindings b : bindings) {
+      final String propertyValue = b.getPropertyValue().getValue();
+      propertyValues.add(propertyValue);
+    }
+
+    return propertyValues;
+  }
+
+  /* see superclass */
+  @Override
+  public List<String> getPropertyValues(final String propertyCode, final Terminology terminology)
+      throws Exception {
+    final String queryPrefix = queryBuilderService.constructPrefix(terminology);
+    final Map<String, String> values =
+        ConceptUtils.asMap("propertyCode", propertyCode, "namedGraph", terminology.getGraph());
+    final String query =
+        queryBuilderService.constructQuery("concept.property.values", terminology, values);
     final String res = restUtils.runSPARQL(queryPrefix + query, getQueryURL());
 
     final ObjectMapper mapper = new ObjectMapper();