+++ src/test/java/gov/nih/nci/evs/api/fhir/FhirR5ValueSetReadSearchTests.java
@@ -275,7 +275,7 @@ class FhirR5ValueSetReadSearchTests {
         UriComponentsBuilder.fromUriString(endpoint)
             .queryParam("_count", "2000")
             .queryParam("_id", "ncit_c100110")
-            // TODO
+            // NOTE  investigate on ticket EVSRESTAPI-699
             // .queryParam("code", "C100110")
             .queryParam("name", "CDISC Questionnaire Terminology")
             .queryParam("title", "ncit")
@@ -384,7 +384,7 @@ class FhirR5ValueSetReadSearchTests {
    */
   @Test
   public void testValueSetSearchWithParameters() throws Exception {
-    // TODO: currently data doesn't have the system data and only subsets have
+    // NOTE: currently data doesn't have the system data and only subsets have
     // the code
     // these tests should be amended when the data is updated