+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
 ## [2.4.0.RELEASE] - 2025-MM-DD
 ### Changed
 - TBD
 
-## [2.3.0.RELEASE] - 2025-11-04
 ### Changed
 - Various minor bug fixes and typo corrections
 - FHIR improvements (resource _history, improved parameter support, POST call support for adhoc value sets, set dates properly on ConceptMap)
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Address CodeQL, dependabot, and trivy reported vulnerability issues
 - Support viewable maps the NCI Thesaurus mappings
 - Bring online support for CTCAE6
-- Improve term form to support attachments in the style of the CDISC template
+- Improve term form to support attachments in the style of the CDISC template (and improve error messages)
 - Better handling of NPO complex definitions
 - Swagger version update